+++ packages/cli/src/command/generate/index.ts
@@ -28,12 +28,12 @@ export default async function (options) {
 
   goldlog('generate', materialConfig);
 
-  const [blocks, components, scaffolds, pages] = await Promise.all(
-    ['block', 'component', 'scaffold', 'page'].map((item) => {
+  const [blocks, components, scaffolds] = await Promise.all(
+    ['block', 'component', 'scaffold'].map((item) => {
       return globMaterials(cwd, item);
     }),
   );
-  const allMaterials = [].concat(blocks).concat(components).concat(scaffolds).concat(pages);
+  const allMaterials = [].concat(blocks).concat(components).concat(scaffolds);
 
   const concurrency = Number(process.env.CONCURRENCY) || 30;
   log.info('Generate:', `generating materials data，total: ${allMaterials.length}，concurrency: ${concurrency}`);
@@ -66,7 +66,6 @@ export default async function (options) {
   const blocksData = [];
   const componentsData = [];
   const scaffoldsData = [];
-  const pagesData = [];
   materialsData.forEach((item) => {
     const { materialType, materialData } = item;
     if (materialType === 'block') {
@@ -75,8 +74,6 @@ export default async function (options) {
       componentsData.push(materialData);
     } else if (materialType === 'scaffold') {
       scaffoldsData.push(materialData);
-    } else if (materialType === 'page') {
-      pagesData.push(materialData);
     }
   });
 
@@ -89,7 +86,6 @@ export default async function (options) {
     blocks: blocksData,
     components: componentsData,
     scaffolds: scaffoldsData,
-    pages: pagesData,
   };
 
   const distFilepath = path.join(cwd, DB_PATH);