+++ extensions/project-creator/CHANGELOG.md
@@ -3,6 +3,10 @@
 
 - feat: update appworks-lab to apptools-lab
 
+## 1.0.8
+
+- chore: update taobao npm registry from `https://registry.npm.taobao.org` to `https://registry.npmmirror.com`
+
 ## 1.0.7
 
 - feat: update appworks-lab to apptools-lab